refactor: components 내 modal 폴더 생성
